requirements.txt
asgiref==3.6.0
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
distro==1.8.0
Django==4.1.5
django-filter==22.1
django-restframework==0.0.1
djangorestframework==3.14.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.2
packaging==23.0
paramiko==3.0.0
psycopg2==2.9.5
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dotenv==0.21.1
pytz==2022.7
PyYAML==5.4.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
texttable==1.6.7
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
websocket-client==0.59.0
